move ds tests to resource_test
